Move code into modules
This commit moves the code out of the main file and into modules which
can be `require`d in Node. I've done this to allow the individual
modules to be tested more easily, and to allow a bundle to be created
using `browserify`.

In order to allow the parser to find the AST functions, I also had to
generate a module for the parser and require that too. This module is
generated using the `pegjs` command line tool

    pegjs ./lib/musicn_parser.js