A simple integration test
Make sure that the system generates 8 seconds of data when passed a
simple score with a TER instruction.

I've had to make the AST object global, as otherwise it isn't in scope
when the parser is generated on-the-fly under Node. I'm sure there's a
better way to do this, but I'm not sure what it is.