Move Score parameters on to Score constructor
A score (in the book I am basing this on) is written with three
parameters in mind:

- rate
- blockSize
- maxAmplitude

The parameters passed to instrument definitions are then in reference to
these units. For example if the maxAmplitude of a score is 2000, then
the loudest note that can be played is 2000. Similarly the frequency is
given in a score in units of (blockSize / rate), so 440Hz when the
blockSize is 511 and the rate is 20000 is 11.242.

These units are not particularly helpful when doing calculations to
generate the sample data, so we scale them as soon as possible into
relative amplitudes (amplitude 1 is maximum) and frequencies (in Hz).